refactor: アイコンモード数字モードのメニューなど
