+++ src/components/atoms/ConfigMenu/MenuButton/index.ts
@@ -1,0 +1,2 @@
+import MenuButton from './MenuButton';
+export default MenuButton;
