+++ src/components/atoms/ConfigMenu/MenuStack/MenuStack.scss
