+++ src/components/atoms/ConfigMenu/MenuStack/index.ts
@@ -1,0 +1,2 @@
+import MenuStack from './MenuStack';
+export default MenuStack;
