--- src/components/menu/ConfigMenu.stories.tsx
@@ -1,8 +1,0 @@
-import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
-import ConfigMenu from './ConfigMenu';
-
-export default {
-  component: ConfigMenu,
-} satisfies ComponentMeta<typeof ConfigMenu>;
-
-export const Default: ComponentStoryObj<typeof ConfigMenu> = {};
